Using different httpClientImpl for twitter streams
MuleHttpClientImpl was not working well
with streams, as it was causing timeouts
when the connection was inactive for some seconds, which
is fine for normal http connections, but not for twitter streams.

Now the connector is using the Mule client for normal twitter
operations and the original twitter htttp client for streaming ops.

The aprroach used is quite hacky, since Twitter4j does not offer
a simple way to inject the http client